refactor(iota-framework-snapshot): split the manifest into a leaf crate
`iota-package-management`'s build script reads the bytecode snapshot
manifest, a JSON file, through `iota-framework-snapshot`. That pulled
`iota-framework` and `iota-types` into its build-dependency tree, and
`iota-package-management` is itself a dependency of `iota-move-build`
and 27 other crates.

The manifest half of the crate has no workspace dependency beyond
`ObjectId`, so it becomes a leaf crate:

- Add `iota-framework-snapshot-manifest` holding the `Snapshot` /
  `SnapshotPackage` types, `manifest_path`, and the manifest load and
  update functions, together with `manifest.json` itself.
- Re-export it from `iota-framework-snapshot` so existing consumers are
  unaffected.
- Replace `SnapshotPackage::from_system_package_metadata`, which the
  orphan rule no longer allows, with the `snapshot_package` function in
  `iota-framework-snapshot`.
- Point the `check_if_manifest_was_modified` test helper at the new
  manifest path, so `check_manifest_against_tomls` keeps running when
  the manifest changes.
