+++ .github/CODEOWNERS
@@ -23,6 +23,7 @@
 /crates/iota-faucet/ @iotaledger/infrastructure
 /crates/iota-framework/ @iotaledger/vm-language
 /crates/iota-framework-snapshot/ @iotaledger/vm-language
+/crates/iota-framework-snapshot-manifest/ @iotaledger/vm-language
 /crates/iota-framework-tests/ @iotaledger/vm-language
 /crates/iota-genesis-builder/ @iotaledger/vm-language
 /crates/iota-genesis-common/ @iotaledger/vm-language
