+++ Cargo.toml
@@ -79,6 +79,7 @@ members = [
   "crates/iota-faucet",
   "crates/iota-framework",
   "crates/iota-framework-snapshot",
+  "crates/iota-framework-snapshot-manifest",
   "crates/iota-framework-tests",
   "crates/iota-genesis-builder",
   "crates/iota-genesis-common",
@@ -390,6 +391,7 @@ iota-execution = { path = "iota-execution" }
 iota-faucet = { path = "crates/iota-faucet" }
 iota-framework = { path = "crates/iota-framework" }
 iota-framework-snapshot = { path = "crates/iota-framework-snapshot" }
+iota-framework-snapshot-manifest = { path = "crates/iota-framework-snapshot-manifest" }
 iota-framework-tests = { path = "crates/iota-framework-tests" }
 iota-genesis-builder = { path = "crates/iota-genesis-builder" }
 iota-genesis-common = { path = "crates/iota-genesis-common" }
