+++ crates/iota-framework-snapshot-manifest/manifest.json
