+++ crates/iota-framework-snapshot/Cargo.toml
@@ -11,13 +11,12 @@ description = "Tool to create a bytecode snapshot of the current iota-framework"
 # external dependencies
 anyhow.workspace = true
 bcs.workspace = true
-serde.workspace = true
-serde_json.workspace = true
 tempfile.workspace = true
 
 # internal dependencies
 bin-version.workspace = true
 iota-framework.workspace = true
+iota-framework-snapshot-manifest.workspace = true
 iota-protocol-config.workspace = true
 iota-sdk-types.workspace = true
 iota-types.workspace = true
